+++ eth/src/test/java/net/consensys/cava/eth/BlockHeaderTest.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package net.consensys.cava.eth.domain;
+package net.consensys.cava.eth;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
