+++ settings.gradle
@@ -4,7 +4,7 @@ include 'concurrent'
 include 'concurrent-coroutines'
 include 'config'
 include 'crypto'
-include 'eth-domain'
+include 'eth'
 include 'eth-reference-tests'
 include 'io'
 include 'junit'
